requirements.txt
anthropic==0.40.0
fastapi==0.115.0
google-api-python-client>=2.100
google-auth>=2.20
httpx==0.27.2
pydantic-settings==2.6.0
pydantic==2.9.2
python-dotenv==1.0.1
uvicorn[standard]==0.32.0
